Connecting to MongoDB: Inside the function, the mongoose.connect method is called with an argument that retrieves the MongoDB URI from the environment variables. This URI is typically stored as MONGO_URI in the environment configuration (.env file) to keep sensitive database credentials secure.
The connect method establishes a connection to the MongoDB instance using the provided URI. The await keyword ensures that the function waits for the connection process to complete before moving forward.

4. Success Message:
If the connection is successful, a message (mongoDB connected.) is printed to the console, indicating that the application has successfully connected to the database.