Type Definitions: DeliveryDetails:
This type defines the structure for the delivery information of an order. It includes:
email: A string for the email of the person receiving the order.
name: A string representing the name of the recipient.
address: The physical address where the order will be delivered.
city: The city of the delivery address.