CartItems:
Defines the structure for items in the cart (the items being ordered):
menuId: A string representing the ID of the menu item being ordered.
name: The name of the menu item.
image: A URL or path to an image of the item.
price: The price of the menu item.
quantity: The number of units of the item being ordered.