Interface Definitions: IMenu:
This is a TypeScript interface that defines the structure of a menu item. It ensures type safety when working with menu documents in the codebase. The fields include:
name: A string representing the name of the menu item.
description: A string describing the menu item.
price: A number representing the price of the menu item.
image: A string containing a URL or path to an image of the menu item.