IMenuDocument:
This interface extends IMenu and Mongoose's Document. It adds additional fields (createdAt and updatedAt) for timestamp tracking. The interface ensures that all menu items adhere to the defined structure and integrates with Mongoose’s document properties.
createdAt: A date representing when the menu item was created.
updatedAt: A date representing when the menu item was last updated.