Menu Schema: The schema defines the structure of each menu document and how it is stored in MongoDB. The following fields are included: name:
A string representing the name of the menu item.
It is required, meaning every menu item must have a name.
description:

A string providing a description of the menu item (e.g., ingredients, preparation style).
This field is also required.
price:

A number representing the cost of the menu item.
This field is required to ensure all menu items have a price.
image:

A string representing the URL or file path of an image for the menu item.
This is required to display images of the food in the application.