Interface & Type Definitions: CartItem Interface:
This interface extends the properties of MenuItem and adds a new property specific to the cart:
quantity: A number representing the quantity of a given menu item in the cart.
Essentially, a CartItem is a MenuItem but with an additional quantity field to track how many of that particular item is in the cart.