Type and Interface Definitions: CheckoutSessionRequest Type:
This type defines the structure of a request to create a checkout session, which typically happens when a user is finalizing their purchase and providing details for an order.

Properties:

cartItems: An array of objects, each representing an item in the cart that the user intends to purchase. Each cart item includes the following fields:

menuId: The unique identifier for the menu item.
name: The name of the menu item.
image: A string representing the image URL or path for the item.
price: The price of the item as a string.
quantity: The quantity of the item being purchased.
deliveryDetails: An object containing the user's delivery information:

name: The name of the person receiving the order.
email: The email address of the customer.
contact: The contact number for the customer.
address: The delivery address.
city: The city for delivery.
country: The country for delivery.
restaurantId: A string representing the unique identifier of the restaurant from which the order is being placed.