Restaurant Type:
Represents a restaurant and its properties.
Properties:
_id: Unique identifier for the restaurant.
user: Identifier for the user (presumably the owner or manager) associated with the restaurant.
restaurantName: Name of the restaurant.
city: The city where the restaurant is located.
country: The country where the restaurant is located.
deliveryTime: Estimated time for delivering orders from this restaurant.
cuisines: An array of strings representing the different cuisines the restaurant offers.
menus: An array of MenuItem objects representing the menu items offered by the restaurant.
imageUrl: URL of the image representing the restaurant.