Event Handlers: changeEventHandler:
Handles changes in the input fields. It updates the input state based on the field being changed. It uses the name attribute of the input to identify which field is being updated and sets the new value accordingly.
loginSubmitHandler:

Handles form submission. It prevents the default form submission behavior.
Validation:
Uses the userSignupSchema to validate the input fields. If validation fails, it sets the errors state with the field-specific error messages.
API Call:
If validation is successful, it attempts to call the signup function from the useUserStore hook with the input data.
On successful signup, it navigates the user to the /verify-email page. If there's an error, it logs it to the console.