Event Handlers:
changeEventHandler: Updates the input state whenever an input field changes. It extracts the name and value of the changed field from the event and updates the state accordingly.
loginSubmitHandler: Handles the form submission:
Validation: Uses userLoginSchema to validate the input data. If validation fails, it updates the errors state with field-specific error messages.
Login Attempt: If validation is successful, it calls the login function from useUserStore with the input data. Upon successful login, it navigates to the home page ("/"). If an error occurs, it logs the error to the console.